requirements.txt
Flask==2.1.2
slackclient==2.9.4
slackeventsapi==3.0.1
gunicorn==20.1.0
python-dotenv